+++ app/src/main/java/com/cargopull/executor_driver/view/MovingToClientFragment.java
@@ -321,7 +321,7 @@ public class MovingToClientFragment extends BaseFragment implements MovingToClie
   }
 
   @Override
-  public void showOrderConditions(@NonNull String routeDistance, int time, int cost) {
+  public void showOrderConditions(@NonNull String routeDistance, int time, long cost) {
 
   }
